fix: tests had been failing due to prevous changes improve build/install process force overwrite is now class default
