+++ .gitignore
@@ -109,3 +109,5 @@ tmp/
 temp/
 
 # End of https://www.gitignore.io/api/node
+dist
+off
