--- jsdoc_conf.json
@@ -1,33 +1,0 @@
-{
-  "tags": {
-    "allowUnknownTags": true,
-    "dictionaries": [
-      "jsdoc",
-      "closure"
-    ]
-  },
-  "source": {
-    "include": [
-      "lib"
-    ],
-    "includePattern": ".+\\.js(doc)?$",
-    "excludePattern": "(^|\\/|\\\\)_"
-  },
-  "opts": {
-    "template": "./node_modules/tidy-jsdoc",
-    "encoding": "utf8",
-    "destination": "./docs/",
-    "recurse": true
-  },
-  "plugins": [
-    "plugins/markdown",
-    "plugins/summarize"
-  ],
-  "templates": {
-    "cleverLinks": false,
-    "monospaceLinks": false
-  },
-  "metadata": {
-    "title": "node_modules"
-  }
-}
