+++ package.json
@@ -1,41 +1,15 @@
 {
   "name": "kvstore",
-  "version": "1.3.0",
+  "version": "2.0.0",
   "description": "Simple Key-Value store with namespaces for simple apps to use",
-  "scripts": {
-    "test": "jest",
-    "lint": "semistandard --verbose | snazzy",
-    "tidy": "semistandard --fix",
-    "generate-docs": "jsdoc --readme README.md -c jsdoc_conf.json",
-    "jsdocs": "jsdoc --readme README.md  -c jsdoc_conf.json"
-  },
+  "type": "commonjs",
   "author": "Kevin Mulholland",
   "license": "ISC",
-  "bin": {
-    "kv": "bin/kv"
-  },
-  "main": "lib/kvstore.js",
   "dependencies": {
     "commander": "^12.0.0",
     "debug": "^4.3.4",
     "dot-json": "^1.3.0",
-    "flat": "=6.0.1",
-    "hash-sum": "^2.0.0",
-    "jsdoc": "^4.0.2",
-    "lockfile": "^1.0.4",
-    "shelljs": "^0.8.5"
-  },
-  "devDependencies": {
-    "jest": "^29.7.0",
-    "path-parse": ">=1.0.7",
-    "semistandard": "^17.0.0",
-    "snazzy": "^9.0.0",
-    "sprintf-js": "^1.1.3",
-    "tidy-jsdoc": "^1.4.1"
-  },
-  "semistandard": {
-    "ignore": [
-      "*test.js"
-    ]
+    "flat": "^6.0.1",
+    "lockfile": "^1.0.4"
   }
 }
